Move volume icon to bottom
